requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
fastapi==0.135.1
fastf1==3.8.1
fonttools==4.61.1
h11==0.16.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
msgpack==1.1.2
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
py4j==0.10.9.9
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
RapidFuzz==3.14.3
requests==2.32.5
requests-cache==1.3.1
scipy==1.17.1
signalrcore==1.0.2
six==1.17.0
starlette==0.52.1
timple==0.1.8
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
google-cloud-bigquery==3.20.1
